Hundle errors on delete category/articles

+++ src/Controller/Back/ArticleController.php
@@ -7,6 +7,7 @@ use App\Repository\ArticleRepository;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\Session\SessionInterface;
 use Symfony\Component\Routing\Annotation\Route;
 use Symfony\Component\String\Slugger\SluggerInterface;
 use Symfony\Component\HttpFoundation\File\Exception\FileException;
@@ -90,10 +91,16 @@ class ArticleController extends AbstractController
     }
 
     #[Route('/{id}', name: 'article_delete', methods: ['POST'])]
-    public function delete(Request $request, Article $article, ArticleRepository $articleRepository): Response
+    public function delete(Request $request, Article $article, ArticleRepository $articleRepository, SessionInterface $session): Response
     {
         if ($this->isCsrfTokenValid('delete'.$article->getId(), $request->request->get('_token'))) {
-            $articleRepository->remove($article, true);
+            $orders = $article->getOrders();
+            if ($orders->isEmpty()) {
+                $articleRepository->remove($article, true);
+                $session->getFlashBag()->add('success', 'Article has been deleted successfully.');
+            } else {
+                $session->getFlashBag()->add('error', 'Article cannot be deleted as it is linked to one or more orders.');
+            }
         }
 
         return $this->redirectToRoute('admin_app_menu', [], Response::HTTP_SEE_OTHER);
